requirements.txt
Flask>=2.0
requests>=2.25
python-telegram-bot==13.15
gunicorn>=20.0
filelock>=3.0
